+++ mobile/android/android-components/components/service/sync-logins/src/main/java/mozilla/components/service/sync/logins/DefaultLoginValidationDelegate.kt
@@ -24,27 +24,103 @@ class DefaultLoginValidationDelegate(
     private val scope: CoroutineScope = CoroutineScope(IO)
 ) : LoginValidationDelegate {
 
-    @Suppress("ComplexMethod") // This method is not actually complex
-    override fun validateCanPersist(login: Login): Deferred<Result> {
+    /**
+     * Compares a [Login] to a passed in list of potential dupes [Login] or queries underlying
+     * storage for potential dupes list of [Login] to determine if it should be updated or created.
+     */
+    override fun shouldUpdateOrCreateAsync(
+        newLogin: Login,
+        potentialDupes: List<Login>?
+    ): Deferred<Result> {
+        return scope.async {
+            val potentialDupesList =
+                potentialDupes ?: storage.value.getPotentialDupesIgnoringUsername(newLogin)
+            val foundLogin = if (potentialDupesList.isEmpty()) {
+                // If list is empty, no records match -> create
+                null
+            } else {
+                // Matching guid, non-blank matching username -> update
+                potentialDupesList.find { it.guid == newLogin.guid && it.username == newLogin.username }
+                // Matching guid, blank username -> update
+                    ?: potentialDupesList.find { it.guid == newLogin.guid && it.username.isEmpty() }
+                    // Matching username -> update
+                    ?: potentialDupesList.find { it.username == newLogin.username }
+                    // Non matching guid, blank username -> update
+                    ?: potentialDupesList.find { it.username.isEmpty() }
+                    // else create
+            }
+            if (foundLogin == null) Result.CanBeCreated else Result.CanBeUpdated(foundLogin)
+        }
+    }
+
+    override fun getPotentialDupesIgnoringUsernameAsync(newLogin: Login): Deferred<List<Login>> {
+        return scope.async {
+            storage.value.getPotentialDupesIgnoringUsername(newLogin)
+        }
+    }
+
+    /**
+     * Queries underlying storage for a new [Login] to determine if should be updated or created,
+     * and can be merged and saved, or any error states.
+     *
+     * @return a [Result] detailing whether we should create or update (and the merged login to update),
+     * else any error states
+     */
+    @Suppress("ComplexMethod")
+    fun ensureValidAsync(login: Login): Deferred<Result> {
         return scope.async {
             try {
-                // We're setting guid=null here in order to trigger de-duping logic properly.
-                // Internally, the library ensures to not dupe records against themselves
-                // (via a `guid <> :guid` check), which means we need to "pretend" this is a new record,
-                // in order for `ensureValid` to actually throw `DUPLICATE_LOGIN`.
-                storage.value.ensureValid(login.copy(guid = null))
-                Result.CanBeCreated
+                var result = shouldUpdateOrCreateAsync(login).await()
+                val loginToCheck = when (result) {
+                    // If we are updating, let's validate the merged login
+                    is Result.CanBeUpdated -> result.foundLogin.mergeWithLogin(login)
+                    is Result.CanBeCreated -> login
+                    else -> login
+                }
+                storage.value.ensureValid(loginToCheck)
+                // Return the merged login to update to avoid doing this merge twice
+                if (result is Result.CanBeUpdated) {
+                    result = Result.CanBeUpdated(loginToCheck)
+                }
+                result
             } catch (e: InvalidRecordException) {
                 when (e.reason) {
-                    InvalidLoginReason.DUPLICATE_LOGIN -> Result.CanBeUpdated
+                    InvalidLoginReason.DUPLICATE_LOGIN -> Result.Error.Duplicate
                     InvalidLoginReason.EMPTY_PASSWORD -> Result.Error.EmptyPassword
                     InvalidLoginReason.EMPTY_ORIGIN -> Result.Error.GeckoError(e)
                     InvalidLoginReason.BOTH_TARGETS -> Result.Error.GeckoError(e)
                     InvalidLoginReason.NO_TARGET -> Result.Error.GeckoError(e)
-                    // TODO in what ways can the login fields be illegal? represent these in the UI
                     InvalidLoginReason.ILLEGAL_FIELD_VALUE -> Result.Error.GeckoError(e)
                 }
             }
+        }
+    }
+
+    companion object {
+        /**
+         * Will use values from [login] if they are 1) non-null and 2) non-empty.  Otherwise, will fall
+         * back to values from [this].
+         */
+        fun Login.mergeWithLogin(login: Login): Login {
+            infix fun String?.orUseExisting(other: String?) =
+                if (this?.isNotEmpty() == true) this else other
+
+            infix fun String?.orUseExisting(other: String) =
+                if (this?.isNotEmpty() == true) this else other
+
+            val origin = login.origin orUseExisting origin
+            val username = login.username orUseExisting username
+            val password = login.password orUseExisting password
+            val httpRealm = login.httpRealm orUseExisting httpRealm
+            val formActionOrigin = login.formActionOrigin orUseExisting formActionOrigin
+
+            return copy(
+                origin = origin,
+                username = username,
+                password = password,
+                httpRealm = httpRealm,
+                formActionOrigin = formActionOrigin
+            )
         }
     }
 }